refactor: AIChatPage 로직 분리 + z카드 랜덤 UI & 해설 공유
