+++ frontend/src/constants/tarotCardNames.ts
@@ -1,0 +1,24 @@
+export const tarotCardNames = [
+  '바보',
+  '마법사',
+  '여사제',
+  '여황제',
+  '황제',
+  '교황',
+  '연인',
+  '전차',
+  '힘',
+  '은둔자',
+  '운명의 수레바퀴',
+  '정의',
+  '매달린 남자',
+  '죽음',
+  '절제',
+  '악마',
+  '탑',
+  '별',
+  '달',
+  '태양',
+  '심판',
+  '세계',
+];
